Update Secure Token Users.sh
Code as it was did not work for me.  I went through and cleaned up a few areas.  I removed the brackets ({}) from the for loop as that was broken.  I also updated some variables with double-quotes (") to prevent globbing/splitting.  I also added a check for greater than 1 entry in each array to prevent a stray comma.